+++ src/main/buffs/Decay.java
@@ -11,7 +11,7 @@ public class Decay extends Buff {
 
     private int damage;
 
-    public Decay(PApplet p, int enId, int damage, int duration, Turret turret) {
+    public Decay(PApplet p, int enId, float damage, int duration, Turret turret) {
         super(p,enId,turret);
         effectDelay = 6; //frames
         effectTimer = p.frameCount + effectDelay;
@@ -19,7 +19,7 @@ public class Decay extends Buff {
         lifeTimer = p.frameCount + lifeDuration;
         particle = "decay";
         name = "decay";
-        this.damage = damage;
+        this.damage = (int) damage;
         this.enId = enId;
         this.turret = turret;
     }
